Consolidate tests related to recommendations endpoint and remove 'proof of concept' test that bypassed the endpoint and made independent connection to Yelp API
